fix(ui): widen UI_INSTANCE_BirthWithThread's pointer params and HUD fields
param_2 (funcThTick), param_5 (pushBuffer), and param_6 (name) are real
pointers carried through retail's int-sized parameters; every one of the
13 call sites truncated them with an explicit `(int)` cast (a function
pointer, a PushBuffer*, and a name string respectively). Widened the
parameters to their real pointee types and dropped the casts -- the name
arguments are `char[]` arrays that decay to pointers without one, the
function pointers now cast to `(void *)` instead of `(int)`.

Also widened sdata->ptrPushBufferUI and sdata->ptrFruitDisp from `int` to
`uintptr_t`: both already had a `should maybe be a pointer` comment, and
ptrPushBufferUI was being stored via `(int)&sdata->pushBuffer_DecalMP` --
already-truncated before its two existing read sites' `(uintptr_t)` casts
ever ran.

+++ include/regionsEXE.h
@@ -3436,10 +3436,10 @@ struct sData
 	// 8008c40c JpnTrial
 	// 8008d858 EurRetail
 	// 800908cc JpnRetail
-	int ptrPushBufferUI;
+	uintptr_t ptrPushBufferUI;
 
 	// 8008d4b8
-	int ptrFruitDisp; // should maybe be `struct Instance*` instead of `int`?
+	uintptr_t ptrFruitDisp; // should maybe be `struct Instance*` instead of `int`?
 
 	// 8008d4bc
 	int menuReadyToPass;
